+++ src/app/shared/list-base/list-base.component.css
@@ -1,0 +1,99 @@
+.list-card {
+  margin: auto;
+  width: fit-content;
+  min-width: 70%;
+  min-height: 80vh;;
+  text-align: center;
+/*   justify-content: center; */
+}
+
+mat-checkbox {
+  --mat-sys-body-medium-size: 10px;
+}
+
+.small-text {
+  font-size: 12px;
+}
+
+.mat-mdc-form-field {
+  margin-left: auto;
+  font-size: 12px;
+  width: 100%;
+}
+
+mat-label {
+  font-size: 12px;
+}
+
+mat-card-header {
+  display: block;
+  justify-content: center;
+  padding-bottom: 8px;
+}
+mat-card-actions {
+  display: block;
+  justify-content: center;
+  padding-bottom: 16px;
+}
+
+ ::ng-deep .mdc-text-field--filled:not(.mdc-text-field--disabled) {
+  background-color: #f9faf4 !important;
+}
+
+.bottom-left {
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  height: 100%;
+  width: 100%;
+  text-align: left;
+}
+
+.bottom-right {
+  width: 100%;
+  text-align: right;
+  padding-right: 16px;
+}
+
+.bottom-right-left-align {
+  width: 100%;
+  text-align: left;
+}
+
+.nav-button-right{
+  width: 100%;
+  text-align: right;
+  padding: 16px 16px 0 0;
+}
+
+.nav-button-left{
+  width: 100%;
+  text-align: left;
+  padding: 16px 0 0 16px ;
+}
+
+.nav-header{
+  width: 100%;
+  text-align: center;
+  padding: 16px 0 16px 0;
+}
+
+mat-button-toggle-group {
+  margin-bottom: 20px;
+}
+
+ mat-card-header {
+  margin: 40px 0 40px 0;
+}
+
+.icon-cursor {
+  cursor: pointer;
+}
+
+.line-through {
+  text-decoration: line-through;
+}
+
+.example-option:focus {
+  background: #d3e8d2;
+}
